Split infra tests into per-topic files, revert conftest.py whitespace
